Script changes.

+++ lab7/bash_args.sh
@@ -1,5 +1,4 @@
 #!/usr/bin/env bash
-#echo $#
 if [[ $# -gt 2 ]] 
 	then
 	echo -e "Invalid usage:\nType $0 or $0 --help to see the proper usages of this script."
@@ -15,10 +14,14 @@ if [[ $# -eq 0 ]] || [[ $1 == "--help" ]]
 	exit 0
 fi
 
-if [[ $# -eq 1 ]]
+if [[ $# -eq 1 ]] && [[ -d "$1" ]] 
 	then 
-	[[ ! -d "$1" ]] && echo "Directory provided as arg1 is not valid." && exit 1
 	inputed_directory="$1"
+	if [[ ! -d "$inputed_directory" ]]
+		then
+		echo "Directory provided as arg1 is not valid."
+		exit 1
+	fi
 	echo "Removing the following files: "
 	find $(realpath "$inputed_directory") -type f -size 0 -print -delete
 	exit 0
@@ -29,7 +32,8 @@ if [[ $# -eq 2 ]] && [[ "$2" =~ ^[0-9]+$ ]] && [[ $2 -ge 1 ]]
 	then
 	inputed_directory="$1"
 	number_of_files="$2"
-	mkdir "$inputed_directory" 
+	[[ -d "$inputed_directory" ]] && /bin/rm -rf "$inputed_directory"
+	mkdir -p "$inputed_directory" 
 	for n in $(seq 1 $number_of_files);
 		do
 		touch "$inputed_directory"/"$n".file
